Move banners into dashboard page component
