+++ ui/src/Pages/Dashboard/Banners/Banner.jsx
@@ -3,11 +3,11 @@ import { useSelector } from "react-redux";
 
 import ProgressBar from "./ProgressBar.jsx";
 import Image from "./Image.jsx";
-import TruncText from "../../Helpers/TruncText.jsx";
-import NewLibraryModal from "../../Modals/NewLibrary/Index";
-import SelectMediaFile from "../../Modals/SelectMediaFile/Index.jsx";
-import SelectMediaFilePlayButton from "../../Modals/SelectMediaFile/Activators/PlayButton.jsx";
-import CircleIcon from "../../assets/Icons/Circle";
+import TruncText from "../../../Helpers/TruncText.jsx";
+import NewLibraryModal from "../../../Modals/NewLibrary/Index";
+import SelectMediaFile from "../../../Modals/SelectMediaFile/Index.jsx";
+import SelectMediaFilePlayButton from "../../../Modals/SelectMediaFile/Activators/PlayButton.jsx";
+import CircleIcon from "../../../assets/Icons/Circle";
 
 import "./Banner.scss";
