+++ ui/src/Pages/Dashboard/Banners/Index.scss
