+++ ui/src/Pages/Dashboard/Banners/ProgressBar.scss
